Training/requirements.txt
kagglehub
PyMuPDF
Pillow
transformers
torch
pytesseract
torchvision
tqdm
requests
openimages
wandb
huggingface_hub